+++ src/hunt/testing/__init__.py
@@ -1,4 +1,12 @@
-from hunt.testing.fakes import EventFake, QueueFake, freeze_time
-from hunt.testing.test_case import HuntTestCase, RefreshDatabase
+from hunt.testing.fakes import EventFake, MailFake, QueueFake, freeze_time
+from hunt.testing.test_case import DatabaseTransactions, HuntTestCase, RefreshDatabase
 
-__all__ = ["EventFake", "HuntTestCase", "QueueFake", "RefreshDatabase", "freeze_time"]
+__all__ = [
+    "DatabaseTransactions",
+    "EventFake",
+    "HuntTestCase",
+    "MailFake",
+    "QueueFake",
+    "RefreshDatabase",
+    "freeze_time",
+]